refactor: move stac item import to EOProduct
